Added a ScreenReaderAbstract with all the common methods and properties from linux, mac and windows screen reader classes
